Replaced treasure chest sprite name with technical placeholder; allowed for negative NPC actor coordinates

+++ SF3Lib/Models/Structs/X1/Town/Npc.cs
@@ -97,21 +97,21 @@ namespace SF3.Models.Structs.X1.Town {
         [TableViewModelColumn(addressField: nameof(_xPosAddr), displayOrder: 6, displayName: "xPos (dec)")]
         [BulkCopy]
         public float ActorX {
-            get => XPos / 65536.0f;
+            get => ((int) XPos) / 65536.0f;
             set => XPos = (uint) value * 65536;
         }
 
         [TableViewModelColumn(addressField: nameof(_yPosAddr), displayOrder: 7, displayName: "yPos (dec)")]
         [BulkCopy]
         public float ActorY {
-            get => YPos / 65536.0f;
+            get => ((int) YPos) / 65536.0f;
             set => YPos = (uint) value * 65536;
         }
 
         [TableViewModelColumn(addressField: nameof(_zPosAddr), displayOrder: 8, displayName: "zPos (dec)")]
         [BulkCopy]
         public float ActorZ {
-            get => ZPos / 65536.0f;
+            get => ((int) ZPos) / 65536.0f;
             set => ZPos = (uint) value * 65536;
         }
